+++ src/shared/utils/create-url-query-params/__test__/createURLQueryParams.test.ts
@@ -1,5 +1,5 @@
 import { faker } from "@faker-js/faker";
-import { createURLQueryParams } from "../createURLQueryParams";
+import { createURLQueryParams } from "../create-url-query-params";
 
 describe("Function createURLQueryParams()", () => {
 	test("Should return url with query params", () => {
@@ -7,10 +7,10 @@ describe("Function createURLQueryParams()", () => {
 		const user_id = faker.datatype.uuid();
 		const department = faker.commerce.department();
 		const queryParams = { user_id, department };
-        
+
 		const url = createURLQueryParams(baseURL, queryParams);
 		const expected = `/contact?user_id=${user_id}&department=${department}`;
-        
+
 		expect(url).toBe(expected);
 	});
 });
