+++ src/shared/utils/create-url-query-params/create-url-query-params.ts
@@ -1,9 +1,12 @@
-import { serializeQueryParams } from "./serializeQueryParams";
+import { serializeQueryParams } from "./serialize-query-params";
 import { ICreateURLQueryParams } from "./types";
 
 const createURLQueryParams: ICreateURLQueryParams = (baseURL, queryParams) => {
 	const queryParamsKeys = Object.keys(queryParams);
-	const concatenatedQueryParams = serializeQueryParams(queryParamsKeys, queryParams);
+	const concatenatedQueryParams = serializeQueryParams(
+		queryParamsKeys,
+		queryParams
+	);
 
 	return `${baseURL}?${concatenatedQueryParams}`;
 };
