+++ src/shared/utils/create-url-query-params/types/index.ts
@@ -1,0 +1,2 @@
+export * as default from "./create-url-query-params.types";
+export * from "./create-url-query-params.types";
