--- src/shared/utils/createURLQueryParams/index.ts
@@ -1,1 +1,0 @@
-export * from "./createURLQueryParams";
