--- src/shared/utils/createURLQueryParams/types/index.ts
@@ -1,2 +1,0 @@
-export * as default from "./createURLQueryParams.types";
-export * from "./createURLQueryParams.types";
